requirements.txt
# JACK Voice Assistant Requirements
# Python 3.12+ recommended

# JACK audio
JACK-Client==0.5.5

# Speech recognition
vosk==0.3.45

# Text-to-speech
piper-tts==1.3.0
onnxruntime==1.17.0

# Audio processing
numpy==1.26.4
scipy==1.16.3

# LLM integration
requests==2.32.5

# Database
psycopg2-binary==2.9.9

# Configuration and utilities
six==1.17.0
